BUGFIX: support long output
If a job generates lots of output, followup jobs cannot properly start.
Errors are like:

  "Task failed with exit code -1:fork/exec /app/flow: argument list too long"

We need to disable storing the task output as environment variable;
otherwise we have quite serious limits how much we can execute.

This seems to be the problem https://stackoverflow.com/a/1078125
"The total size of all the environment variables put together is limited
at execve() time. See "Limits on size of arguments and environment" at
https://man7.org/linux/man-pages/man2/execve.2.html for more information.

On some systems here, this is around 2 MB (not enough!)